updating all service and controller

+++ srcs/backend/src/custom/custom.service.ts
@@ -1,5 +1,5 @@
-import { Injectable } from '@nestjs/common';
-import { PrismaService } from 'src/prisma/prisma.service';
+import { Injectable, NotFoundException } from '@nestjs/common';
+import { PrismaService } from 'src/prisma/prisma.service'; // Assurez-vous d'utiliser le chemin correct
 import { Custom, Prisma } from '@prisma/client';
 
 @Injectable()
@@ -10,8 +10,16 @@ export class CustomService {
     return this.prisma.custom.create({ data });
   }
 
-  async getCustomById(id: number): Promise<Custom | null> {
-    return this.prisma.custom.findUnique({ where: { id } });
+  async getCustomById(id: number): Promise<Custom> {
+    const custom = await this.prisma.custom.findUnique({
+      where: {
+        id,
+      },
+    });
+    if (!custom) {
+      throw new NotFoundException('Custom item not found');
+    }
+    return custom;
   }
 
   async getAllCustoms(): Promise<Custom[]> {
@@ -19,6 +27,10 @@ export class CustomService {
   }
 
   async updateCustom(id: number, data: Prisma.CustomUpdateInput): Promise<Custom | null> {
+    const existingCustom = await this.prisma.custom.findUnique({ where: { id } });
+    if (!existingCustom) {
+      throw new NotFoundException(`Custom item with ID ${id} not found`);
+    }
     return this.prisma.custom.update({
       where: { id },
       data,
@@ -26,6 +38,10 @@ export class CustomService {
   }
 
   async deleteCustom(id: number): Promise<Custom | null> {
+    const existingCustom = await this.prisma.custom.findUnique({ where: { id } });
+    if (!existingCustom) {
+      throw new NotFoundException(`Custom item with ID ${id} not found`);
+    }
     return this.prisma.custom.delete({ where: { id } });
   }
 }
