+++ srcs/backend/src/friendship/friendship.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service'; // Assurez-vous d'utiliser le chemin correct
 import { Friendship, Prisma } from '@prisma/client';
 
@@ -10,8 +10,16 @@ export class FriendshipService {
     return this.prisma.friendship.create({ data });
   }
 
-  async getFriendshipById(id: number): Promise<Friendship | null> {
-    return this.prisma.friendship.findUnique({ where: { id } });
+  async getFriendshipById(id: number): Promise<Friendship> {
+    const friendship = await this.prisma.friendship.findUnique({
+      where: {
+        id,
+      },
+    });
+    if (!friendship) {
+      throw new NotFoundException('Friendship not found');
+    }
+    return friendship;
   }
 
   async getAllFriendships(): Promise<Friendship[]> {
@@ -19,6 +27,10 @@ export class FriendshipService {
   }
 
   async updateFriendship(id: number, data: Prisma.FriendshipUpdateInput): Promise<Friendship | null> {
+    const existingFriendship = await this.prisma.friendship.findUnique({ where: { id } });
+    if (!existingFriendship) {
+      throw new NotFoundException(`Friendship with ID ${id} not found`);
+    }
     return this.prisma.friendship.update({
       where: { id },
       data,
@@ -26,6 +38,10 @@ export class FriendshipService {
   }
 
   async deleteFriendship(id: number): Promise<Friendship | null> {
+    const existingFriendship = await this.prisma.friendship.findUnique({ where: { id } });
+    if (!existingFriendship) {
+      throw new NotFoundException(`Friendship with ID ${id} not found`);
+    }
     return this.prisma.friendship.delete({ where: { id } });
   }
 }
