+++ srcs/backend/src/member/member.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service'; // Assurez-vous d'utiliser le chemin correct
 import { Member, Prisma } from '@prisma/client';
 
@@ -10,8 +10,16 @@ export class MemberService {
     return this.prisma.member.create({ data });
   }
 
-  async getMemberById(id: number): Promise<Member | null> {
-    return this.prisma.member.findUnique({ where: { id } });
+  async getMemberById(id: number): Promise<Member> {
+    const member = await this.prisma.member.findUnique({
+      where: {
+        id,
+      },
+    });
+    if (!member) {
+      throw new NotFoundException('Member not found');
+    }
+    return member;
   }
 
   async getAllMembers(): Promise<Member[]> {
@@ -19,6 +27,10 @@ export class MemberService {
   }
 
   async updateMember(id: number, data: Prisma.MemberUpdateInput): Promise<Member | null> {
+    const existingMember = await this.prisma.member.findUnique({ where: { id } });
+    if (!existingMember) {
+      throw new NotFoundException(`Member with ID ${id} not found`);
+    }
     return this.prisma.member.update({
       where: { id },
       data,
@@ -26,6 +38,10 @@ export class MemberService {
   }
 
   async deleteMember(id: number): Promise<Member | null> {
+    const existingMember = await this.prisma.member.findUnique({ where: { id } });
+    if (!existingMember) {
+      throw new NotFoundException(`Member with ID ${id} not found`);
+    }
     return this.prisma.member.delete({ where: { id } });
   }
 }
